Fix: Review changes for simplicity
